Refactor cpu opcodes to not return an optional event emission and have the emit opcode just do that itself
